plasmid function as module
